Revise the network to take into consideration the current values

+++ utils.py
@@ -34,4 +34,4 @@ def save_or_show_plot(file_nm: str, save: bool):
 
 
 def numpy_to_tensor(x):
-    return torch.from_numpy(x).to(device)
+    return torch.from_numpy(x).type(torch.FloatTensor).to(device)
